Refactor AiTrip TripService into focused service classes
